requirements.txt
beautifulsoup4==4.6.0
certifi==2017.11.5
chardet==3.0.4
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11
django-cors-headers==2.1.0
djangorestframework==3.7.3
gunicorn==19.7.1
idna==2.6
psycopg2==2.7.3.2
pytz==2017.3
requests==2.18.4
static3==0.7.0
urllib3==1.22
